Split net api and web3 api into their own modules
