+++ client/rpc-core/src/net.rs
@@ -17,9 +17,11 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 //! Net rpc interface.
-use crate::types::PeerCount;
+
 use jsonrpc_core::Result;
 use jsonrpc_derive::rpc;
+
+use crate::types::PeerCount;
 
 pub use rpc_impl_NetApi::gen_server::NetApi as NetApiServer;
