+++ client/rpc/Cargo.toml
@@ -20,22 +20,22 @@ jsonrpc-core-client = "18.0"
 jsonrpc-pubsub = "18.0"
 rand = "0.8"
 rlp = "0.5"
-sha3 = "0.10"
 tokio = { version = "1.14", features = [ "sync" ] }
 
 codec = { package = "parity-scale-codec", version = "2.0.0" }
-sp-api = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-blockchain = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-block-builder = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-api = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-blockchain = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-block-builder = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-core = { version = "5.0.0", git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-io = { version = "5.0.0",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-runtime = { version = "5.0.0",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-storage = { version = "5.0.0",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-client-api = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-network = { version = "0.10.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-rpc = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-service = { version = "0.10.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-transaction-pool = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
-sc-transaction-pool-api = { version = "4.0.0-dev",  git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-runtime = { version = "5.0.0", git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-storage = { version = "5.0.0", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-client-api = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-network = { version = "0.10.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-rpc = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-service = { version = "0.10.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-transaction-pool = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
+sc-transaction-pool-api = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate", branch = "master" }
 
 fp-rpc = { version = "3.0.0-dev", path = "../../primitives/rpc" }
 fp-storage = { version = "2.0.0-dev", path = "../../primitives/storage" }
